+++ src/UnderGUI/Utility/__init__.py
@@ -10,6 +10,38 @@ class Range:
         self.y1 = y1
         self.x2 = x2
         self.y2 = y2
+        
+    def __eq__(self, other):
+        return self.x1 == other.x1 and self.y1 == other.y1 and self.x2 == other.x2 and self.y2 == other.y2
+        
+    def __ne__(self, other):
+        return self.x1 != other.x1 or self.y1 != other.y1 or self.x2 != other.x2 or self.y2 != other.y2
+        
+    def __mul__(self, other):
+        if isinstance(other, Range):
+            return Range(self.x1 * other.x1, self.y1 * other.y1, self.x2 * other.x2, self.y2 * other.y2)
+        return Range(self.x1 * other, self.y1 * other, self.x2 * other, self.y2 * other)
+        
+    def __truediv__(self, other):
+        if isinstance(other, Range):
+            return Range(self.x1 / other.x1, self.y1 / other.y1, self.x2 / other.x2, self.y2 / other.y2)
+        return Range(self.x1 / other, self.y1 / other, self.x2 / other, self.y2 / other)
+        
+    def __floordiv__(self, other):
+        print("uuu")
+        if isinstance(other, Range):
+            return Range(self.x1 / other.x1, self.y1 / other.y1, self.x2 / other.x2, self.y2 / other.y2)
+        return Range(self.x1 / other, self.y1 / other, self.x2 / other, self.y2 / other)
+        
+    def __add__(self, other):
+        if isinstance(other, Range):
+            return Range(self.x1 + other.x1, self.y1 + other.y1, self.x2 + other.x2, self.y2 + other.y2)
+        return Range(self.x1 + other, self.y1 + other, self.x2 + other, self.y2 + other)
+        
+    def __sub__(self, other):
+        if isinstance(other, Range):
+            return Range(self.x1 - other.x1, self.y1 - other.y1, self.x2 - other.x2, self.y2 - other.y2)
+        return Range(self.x1 - other, self.y1 - other, self.x2 - other, self.y2 - other)
 
 class ImageInfo:
     def __init__(self, data = None, width = 0, height = 0, err_msg = None):